feat: componentize role edit form (wip)
